♻️ Refactor: manage 도메인 타이핑 마무리
